Move SpeciesStore.acceptProblemSuggestion to SpeciesService
In preparation for adding more logic to species edits, move the code that
accepts suggested edits to SpeciesService so it will be able to access
other store classes. No change in functionality here, though there is a
new test to verify the happy-path behavior.
