+++ src/main/kotlin/com/terraformation/backend/species/SpeciesService.kt
@@ -1,8 +1,14 @@
 package com.terraformation.backend.species
 
 import com.terraformation.backend.customer.model.requirePermissions
+import com.terraformation.backend.db.ScientificNameExistsException
 import com.terraformation.backend.db.SpeciesInUseException
+import com.terraformation.backend.db.SpeciesProblemHasNoSuggestionException
+import com.terraformation.backend.db.SpeciesProblemNotFoundException
 import com.terraformation.backend.db.default_schema.SpeciesId
+import com.terraformation.backend.db.default_schema.SpeciesProblemField
+import com.terraformation.backend.db.default_schema.SpeciesProblemId
+import com.terraformation.backend.db.default_schema.SpeciesProblemType
 import com.terraformation.backend.species.db.SpeciesChecker
 import com.terraformation.backend.species.db.SpeciesStore
 import com.terraformation.backend.species.event.SpeciesEditedEvent
@@ -11,6 +17,7 @@ import com.terraformation.backend.species.model.NewSpeciesModel
 import jakarta.inject.Named
 import org.jooq.DSLContext
 import org.springframework.context.ApplicationEventPublisher
+import org.springframework.dao.DuplicateKeyException
 
 @Named
 class SpeciesService(
@@ -51,5 +58,41 @@ class SpeciesService(
     }
 
     speciesStore.deleteSpecies(speciesId)
+  }
+
+  fun acceptProblemSuggestion(problemId: SpeciesProblemId): ExistingSpeciesModel {
+    val problem = speciesStore.fetchProblemById(problemId)
+    val speciesId = problem.speciesId ?: throw SpeciesProblemNotFoundException(problemId)
+    val existingSpecies = speciesStore.fetchSpeciesById(speciesId)
+
+    val fieldId = problem.fieldId ?: throw IllegalStateException("Species problem had no field")
+    val typeId = problem.typeId ?: throw IllegalStateException("Species problem had no type")
+
+    val correctedSpecies =
+        when (typeId) {
+          SpeciesProblemType.NameNotFound -> throw SpeciesProblemHasNoSuggestionException(problemId)
+          SpeciesProblemType.NameIsSynonym,
+          SpeciesProblemType.NameMisspelled -> {
+            // Only one field defined right now but use a "when" so this will break the build if
+            // we add a second field and forget to handle it here.
+            when (fieldId) {
+              SpeciesProblemField.ScientificName ->
+                  existingSpecies.copy(scientificName = problem.suggestedValue!!)
+            }
+          }
+        }
+
+    return try {
+      dslContext.transactionResult { _ ->
+        speciesStore.deleteProblem(problemId)
+        speciesStore.updateSpecies(correctedSpecies)
+      }
+    } catch (e: DuplicateKeyException) {
+      if (fieldId == SpeciesProblemField.ScientificName) {
+        throw ScientificNameExistsException(problem.suggestedValue)
+      } else {
+        throw e
+      }
+    }
   }
 }
